Add files via upload
I use this code specifically for pitching mechanics for interpretation. The main processing tool is V3D, and this works off of the XML file that is provided, however; it should work with any XML file. You just need to modify how it parses for your specific needs. This is a very simplified version of the code and can be added to until it meets your specific needs. 
This code was written on and for Jupyter files specifically so if you're having any trouble (which there seems to be some formatting issues in plain python files) try switching over to Jupyter.